+++ be/src/main/java/kr/codesquad/secondhand/api/product/dto/response/ProductSlicesResponse.java
@@ -1,4 +1,4 @@
-package kr.codesquad.secondhand.api.product.dto;
+package kr.codesquad.secondhand.api.product.dto.response;
 
 import java.net.URL;
 import java.sql.Timestamp;
