+++ be/src/main/java/kr/codesquad/secondhand/api/product/service/ProductFacadeService.java
@@ -11,10 +11,10 @@ import kr.codesquad.secondhand.api.product.domain.Product;
 import kr.codesquad.secondhand.api.product.domain.ProductImage;
 import kr.codesquad.secondhand.api.product.domain.ProductStats;
 import kr.codesquad.secondhand.api.product.domain.ProductStatus;
-import kr.codesquad.secondhand.api.product.dto.ProductCreateRequest;
-import kr.codesquad.secondhand.api.product.dto.ProductCreateResponse;
-import kr.codesquad.secondhand.api.product.dto.ProductReadResponse;
-import kr.codesquad.secondhand.api.product.dto.ProductUpdateRequest;
+import kr.codesquad.secondhand.api.product.dto.request.ProductCreateRequest;
+import kr.codesquad.secondhand.api.product.dto.request.ProductUpdateRequest;
+import kr.codesquad.secondhand.api.product.dto.response.ProductCreateResponse;
+import kr.codesquad.secondhand.api.product.dto.response.ProductReadResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
